backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
python-calamine==0.2.0
resend>=2.0.0
python-multipart==0.0.12
celery[redis]==5.4.0
redis==5.2.0
